==> notebooks/01_data_cleaning.ipynb <==
## 01 - Data Cleaning

*Get a general overview of the data, clean if needed, and prepare for exploration*

from pathlib import Path
import pandas as pd
from IPython.display import display

project_root = Path.cwd().parent
raw_data = project_root / 'data' / 'raw'

df = pd.read_csv(raw_data / 'walsoft_semi_categorized_phone_dataset.csv')
df_categories = pd.read_csv(raw_data / 'categories.csv')

display(df_categories)

display(df)

Eish! We see an annoying hyphen in the *day-of_week* column, lets fix it because i love standarized stuff

df = df.rename(columns={'day-of_week': 'day_of_week'})

df.describe()

- Majority of calls were made in 2022
- Longest call lasted 7200 seconds (2 hours), shortest 1 second, with a mean of 103 seconds

df.info()


We see column *date_stamp* are of datatype object, let's change *date_stamp* to be of datetime, and add an *hour* column of type integer

df['date_stamp'] = pd.to_datetime(df['date_stamp'])
df['hour'] = pd.to_datetime(df['time'], format='%H:%M:%S').dt.hour

Let's add a boolean column to flag whether a call was made during business hours. We'll define business hours as 8am to 17pm (5pm), Monday through Friday

df['is_business_hours'] = (
    (df['hour'] >= 8) &
    (df['hour'] < 17) &
    (df['day_of_week'].isin(['Monday', 'Tuesday', 'Wednesday', 'Thursday', 'Friday']))
)

missing = df.isnull().sum()
print(missing)

duplicates = df.duplicated().sum()
print(duplicates)

A responsible kaggle user has uploaded this dataset, no missing values or duplicates

df.info()

Great, data is now clean and ready to be added to our /data/processed folder

df.to_csv(project_root / 'data' / 'processed' / 'calls_cleaned.csv', index=False)
